--- templates/thank-you.html
@@ -1,82 +1,0 @@
-{% extends 'index.html' %}
-{% load static %}
-
-<!doctype html>
-<html class="no-js" lang="ru">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="x-ua-compatible" content="ie=edge">
-        <title>{% block title %}Спасибо - {{ block.super }}{% endblock %}</title>
-        <meta name="description" content="">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        
-        <!-- favicon
-		============================================ -->		
-        <link rel="shortcut icon" type="image/x-icon" href="{% static 'img/favicon.ico' %}">
-		<!-- google fonts -->
-		<link href='https://fonts.googleapis.com/css?family=Open+Sans:400,400italic,600,600italic,700,300italic' rel='stylesheet' type='text/css'>
-		<!-- all css here -->
-		<!-- bootstrap v3.3.6 css -->
-        <link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
-		<!-- animate css -->
-        <link rel="stylesheet" href="{% static 'css/animate.css' %}">
-		<!-- pe-icon-7-stroke -->
-		<link rel="stylesheet" href="{% static 'css/pe-icon-7-stroke.min.css' %}">
-		<!-- jquery-ui.min css -->
-        <link rel="stylesheet" href="{% static 'css/jquery-ui.min.css' %}">
-        <!-- Image Zoom CSS
-		============================================ -->
-        <link rel="stylesheet" href="{% static 'css/img-zoom/jquery.simpleLens.css' %}">
-		<!-- meanmenu css -->
-        <link rel="stylesheet" href="{% static 'css/meanmenu.min.css' %}">
-		<!-- nivo slider CSS
-		============================================ -->
-		<link rel="stylesheet" href="{% static 'lib/css/nivo-slider.css' %}" type="text/css" />
-		<link rel="stylesheet" href="{% static 'lib/css/preview.css' %}" type="text/css" media="screen" />
-		<!-- owl.carousel css -->
-        <link rel="stylesheet" href="{% static 'css/owl.carousel.css' %}">
-		<!-- font-awesome css -->
-        <link rel="stylesheet" href="{% static 'css/font-awesome.min.css' %}">
-		<!-- style css -->
-		<link rel="stylesheet" href="{% static 'style.css' %}">
-		<!-- responsive css -->
-        <link rel="stylesheet" href="{% static 'css/responsive.css' %}">
-		<!-- modernizr css -->
-        <script src="{% static 'js/vendor/modernizr-2.8.3.min.js ' %}"></script>
-    </head>
-    <body>
-        <!-- header section start -->
-		<div class="message-area">
-            <div class="container">
-                <div class="row">
-                    <div class="col-md-12 text-center">
-                        <div class="check">
-                            <h2>Message send successfully</h2>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-		<!-- all js here -->
-		<!-- jquery latest version -->
-        <script src="{% static 'js/vendor/jquery-1.12.0.min.js' %}"></script>
-		<!-- bootstrap js -->
-        <script src="{% static 'js/bootstrap.min.js' %}"></script>
-		<!-- owl.carousel js -->
-        <script src="{% static 'js/owl.carousel.min.js' %}"></script>
-		<!-- meanmenu js -->
-        <script src="{% static 'js/jquery.meanmenu.js' %}"></script>
-		<!-- countdown JS -->
-        <script src="{% static 'js/countdown.js' %}"></script>
-		<!-- nivo.slider JS -->
-        <script src="{% static 'js/jquery.nivo.slider.pack.js' %}"></script>
-		<!-- jquery-ui js -->
-        <script src="{% static 'js/jquery-ui.min.js' %}"></script>
-		<!-- sticky js -->
-        <script src="{% static 'js/sticky.js' %}"></script>
-		<!-- plugins js -->
-        <script src="{% static 'js/plugins.js' %}"></script>
-		<!-- main js -->
-        <script src="{% static 'js/main.js' %}"></script>
-    </body>
-</html>
