+++ user_accounts/admin.py
@@ -1,9 +1,9 @@
-from django.contrib import admin
-from .models import User, Order
+# from django.contrib import admin
+# from .models import User, Order
 
 
 
-admin.site.register(User)
-admin.site.register(Order)
+# admin.site.register(User)
+# admin.site.register(Order)
