+++ user_accounts/forms.py
@@ -1,19 +1,19 @@
-from django import forms
-from django.contrib.auth.forms import UserCreationForm
+# from django import forms
+# from django.contrib.auth.forms import UserCreationForm
 
-from .models import User
-
-
-
-class SignUpForm(UserCreationForm):
-
-    class Meta:
-        model = User
-        fields = ('email', 'password', 'password2')
+# from .models import User
 
 
-class UserProfileForm(forms.ModelForm):
+
+# class SignUpForm(UserCreationForm):
+
+#     class Meta:
+#         model = User
+#         fields = ('email', 'password', 'password2')
+
+
+# class UserProfileForm(forms.ModelForm):
     
-    class Meta:
-        model = User
-        fields = ('first_name', 'last_name', 'phone')
+#     class Meta:
+#         model = User
+#         fields = ('first_name', 'last_name', 'phone')
